% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot Diagnostics for an Rttest Object}
\usage{
\method{plot}{Rttest}(
  x,
  main = "Comparison of Value Distripution per Sample",
  sub = NULL,
  poplab = "Population",
  vlab = "Value",
  ...
)
}
\arguments{
\item{x}{Rttest object.}

\item{main}{main title for the plot.}

\item{sub}{sub-title for the plot.}

\item{poplab}{a label to denote what population the samples were taken from
(default: "Population")}

\item{vlab}{a label to denote the values of the samples (default: "Value")}

\item{...}{unused}
}
\description{
A customizable box plot of the two samples contained within the Rttest object
using ggplot2
}
\examples{
x <- rnorm(30,5,2)
y <- rnorm(40,3,2)
a <- 0.1
obj <- Rttest(x, y, a)

plot(obj)
plot(obj, main="My Custom Title", sub="Which also has a sub-title",
     poplab="My Population", vlab="My Value Type")

}
