% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print.Rttest}
\usage{
\method{print}{Rttest}(x, width = NULL, height = 400, ...)
}
\arguments{
\item{x}{Rttest object}

\item{width}{The width in pixels(px). (defualt: NULL)}

\item{height}{The height in pixels(px). This can be set to NULL for and
unbounded height. (defualt: 400)}

\item{...}{unused}
}
\description{
Displays the 2 sample using kable.
}
