% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rttest.R
\name{Rttest}
\alias{Rttest}
\title{Rttest}
\usage{
Rttest(x, y, paired = F, alpha = 0.05)
}
\arguments{
\item{x}{First sample vector}

\item{y}{Second sample vector}

\item{paired}{Whether the samples are paired}

\item{alpha}{determines range of the confidence interval's distance from 100\%}
}
\value{
An Rttest class obj containing a list comprised of a data.frame of
the given samples (x,y), alpha value, confidence interval of type "conf.level",
and p value.
}
\description{
This function preforms several types of t-tests tailored towards user input.
}
\examples{
x <-rnorm(30,5,2)
y<- rnorm(40, 3,2)
alpha <- 0.1
obj <- Rttest(x = x, y = y, alpha = alpha)

}
