% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDemo.R
\name{shinyDemo}
\alias{shinyDemo}
\title{shinyDemo}
\usage{
shinyDemo()
}
\value{
**TODO**
}
\description{
This will help illostrate the functionality of the package utilizing shiny.
}
\examples{
\dontrun{shinyDemo()}
}
