% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print.Rttest}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{Rttest class object}

\item{...}{unused}
}
\description{
Displays the 2 sample using kable.
}
